Fix null pointer exception because of deleting all GuiDevices
When updating the gui because of changes in the api all GuiDevices are deleted and redrawn.
But if the user selected a GuiDevice & Feature before to connect it the pointer points to the old deleted GuiDevice.
This change tries to rediscover the feature on the "new" GuiDevices to point to the right device again.